Initial separation and various bug fixes
All code changed was marked with TODO

Indexes were changed from int to size_t

Removed old master file

Added Visual studio project and .gitignore

CMake needs updating
