+++ AutoCalculus/AutoCalculus.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2027
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AutoCalculus", "AutoCalculus.vcxproj", "{F266F3C1-76A1-4C7A-BD17-5138021F170C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Debug|x64.ActiveCfg = Debug|x64
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Debug|x64.Build.0 = Debug|x64
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Debug|x86.ActiveCfg = Debug|Win32
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Debug|x86.Build.0 = Debug|Win32
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Release|x64.ActiveCfg = Release|x64
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Release|x64.Build.0 = Release|x64
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Release|x86.ActiveCfg = Release|Win32
+		{F266F3C1-76A1-4C7A-BD17-5138021F170C}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8D860D26-715E-43B6-8EB1-6B4F0E2C9CC9}
+	EndGlobalSection
+EndGlobal
